Add error handling
This commit should sophisticate the use of error in this package. The approach is using two primary error variants one for the display to external users and one for internal logs. Also there is a fallback variant if none apply for example when converting another error into an internal one without an addition message.